python/wordcloud/requirements.txt
cycler==0.10.0

image==1.5.5
matplotlib==1.5.3
numpy==1.11.2
Pillow==3.4.2
pkg-resources==0.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
wordcloud==1.2.1
cycler==0.10.0
Django==1.10.3
image==1.5.5
matplotlib==1.5.3
numpy==1.11.2
Pillow==3.4.2
pkg-resources==0.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
wordcloud==1.2.1
